requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Uploads
httpie
psycopg2-binary
passlib
python-dotenv
marshmallow
webargs